all dataset with ensamble bug fixing

+++ mnist/ensemble_learning.py
@@ -71,9 +71,9 @@ def augmentation(source_path, destination_path, classes_dir, output_dir, dataset
             p = Augmentor.Pipeline(source_path + class_dir)
             p.crop_random(probability=1, percentage_area=0.8)
             p.resize(probability=1.0, width=28, height=28)
-            p.sample(sample_number/2)
+            p.sample(int(sample_number/2))
             p.flip_left_right(probability=1.0)
-            p.sample(sample_number/2)
+            p.sample(int(sample_number/2))
         elif technique_determination[class_dir] == 'elastic':
             p = Augmentor.Pipeline(source_path + class_dir)
             p.random_distortion(probability=1, magnitude=2, grid_height=4, grid_width=4)
